logica para validacao de regras de negocio 2

+++ src/main/java/io/github/caiquealves/libraryapi/service/AutorService.java
@@ -1,8 +1,11 @@
 package io.github.caiquealves.libraryapi.service;
 
+import io.github.caiquealves.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.caiquealves.libraryapi.model.Autor;
 import io.github.caiquealves.libraryapi.repository.AutorRepository;
+import io.github.caiquealves.libraryapi.repository.LivroRepository;
 import io.github.caiquealves.libraryapi.validador.AutorValidator;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,10 +16,12 @@ import java.util.UUID;
 public class AutorService {
     private final AutorRepository repository;
     private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository, AutorValidator validator) {
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
         this.repository = repository;
         this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -37,6 +42,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Autor possui livro cadastrado");
+        }
         repository.delete(autor);
     }
 
@@ -50,5 +58,9 @@ public class AutorService {
         }
         return repository.findAll();
 
+    }
+
+    public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
